+++ tests/solidity_test_helpers/MaliciousMarket.sol
@@ -14,11 +14,11 @@ contract MaliciousMarket {
         victimMarket = _market;
     }
 
-    function getShareToken(uint8 _outcome)  public view returns (IShareToken) {
+    function getShareToken(uint256 _outcome)  public view returns (IShareToken) {
         return victimMarket.getShareToken(_outcome);
     }
 
-    function getNumberOfOutcomes() public view returns (uint8) {
+    function getNumberOfOutcomes() public view returns (uint256) {
         return victimMarket.getNumberOfOutcomes();
     }
